requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyPDF2==3.0.1
